+++ src/cpu_lib/managers/getters/__init__.py
@@ -1,3 +1,4 @@
 """Exposed getters classes and methods."""
 from cpu_lib.managers.getters.temperature_getter import TemperatureGetter
 from cpu_lib.managers.getters.use_getter import UseGetter
+from cpu_lib.managers.getters.information_getter import InformationGetter